+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # 21 March 2025
 
-- Added an example to get owner email from ZA Domains
+- Changed namespace from eugenevdm to eugenefvdm
 - Added this change log
+- Added an example to get owner email from ZA Domains
+